refactor(input): apply maintainer review feedback
- extract the duplicated shift/clip edit-adjustment logic into one
  shared unit per platform (RangeEditAdjustment / ENRMRangeEditAdjustment);
  FormattingStore and BlockStore now delegate to it on both platforms
- convert FormattingRange and BlockRange from data classes to plain
  classes: their bounds mutate as edits shift ranges, so value-based
  equals/hashCode would silently break set/map lookups
- correct the applyBlockRanges seeding comment: re-claimed paragraphs
  read the default style after the reset pass; seeding matters only for
  newly claimed paragraphs
- drop comments that narrate what the adjacent code already says
  (InputEventEmitter, MarkdownSerializer on both platforms,
  EnrichedMarkdownTextInput)
